Reserva fet

+++ BackEnd/src/main/java/cat/copernic/ymelero/EntreBicis/entity/Recompensa.java
@@ -33,7 +33,7 @@ public class Recompensa {
 
     private String observacions;
 
-    private Double punts;
+    private Integer punts;
 
     @Enumerated(EnumType.STRING)
     private EstatRecompensa estat;
