+++ FrontEnd/app/src/main/java/cat/copernic/ymelero/entrebicis/recompenses/domain/RecUseCases.kt
@@ -6,7 +6,7 @@ class RecUseCases (private val recRepository: RecRepository) {
     suspend fun getRecompensesDisponibles() = recRepository.getRecompensesDisponibles()
     suspend fun getRecompensesPropies(email: String) = recRepository.getRecompensesPropies(email)
     suspend fun getRecompensaPerId(id: Long) = recRepository.getRecompensaPerId(id)
-    suspend fun reservarRecompensa(recompensaId: Long, email: String, saldo: Long) =
+    suspend fun reservarRecompensa(recompensaId: Long, email: String, saldo: Int) =
         recRepository.reservarRecompensa(recompensaId, email, saldo)
 
 }
